tests: Move classes to `Fixture` dir (#130)
